Rename annotation support module and packages
