+++ instrumentation-api-annotation-support/src/main/java/io/opentelemetry/instrumentation/api/annotation/support/ParameterizedClass.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.annotation.support;
+package io.opentelemetry.instrumentation.api.annotation.support;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
