+++ instrumentation-api-annotation-support/src/test/groovy/io/opentelemetry/instrumentation/api/annotation/support/ParameterizedClassTest.groovy
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.annotation.support
+package io.opentelemetry.instrumentation.api.annotation.support
 
 import spock.lang.Shared
 import spock.lang.Specification
