+++ instrumentation/spring/spring-boot-autoconfigure/build.gradle.kts
@@ -8,7 +8,7 @@ group = "io.opentelemetry.instrumentation"
 val versions: Map<String, String> by project
 
 dependencies {
-  implementation(project(":instrumentation-annotation-support"))
+  implementation(project(":instrumentation-api-annotation-support"))
 
   implementation("org.springframework.boot:spring-boot-autoconfigure:${versions["org.springframework.boot"]}")
   annotationProcessor("org.springframework.boot:spring-boot-autoconfigure-processor:${versions["org.springframework.boot"]}")
@@ -51,7 +51,7 @@ dependencies {
   testImplementation("io.opentelemetry:opentelemetry-exporter-zipkin")
   testImplementation("io.grpc:grpc-api:1.30.2")
   testImplementation("io.grpc:grpc-netty-shaded:1.30.2")
-  testImplementation(project(":instrumentation-annotation-support"))
+  testImplementation(project(":instrumentation-api-annotation-support"))
 }
 
 tasks.compileTestJava {
