+++ instrumentation/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/aspects/WithSpanAspectAttributeBinder.java
@@ -6,8 +6,8 @@
 package io.opentelemetry.instrumentation.spring.autoconfigure.aspects;
 
 import io.opentelemetry.extension.annotations.SpanAttribute;
-import io.opentelemetry.instrumentation.annotation.support.AttributeBindings;
-import io.opentelemetry.instrumentation.annotation.support.BaseAttributeBinder;
+import io.opentelemetry.instrumentation.api.annotation.support.AttributeBindings;
+import io.opentelemetry.instrumentation.api.annotation.support.BaseAttributeBinder;
 import io.opentelemetry.instrumentation.api.caching.Cache;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
